Thanks, Captain-S0L0 ! Table resizes are better.

+++ src/main/java/com/github/tommyettinger/ds/Utilities.java
@@ -187,7 +187,7 @@ public final class Utilities {
 		if (capacity < 0) {
 			throw new IllegalArgumentException("capacity must be >= 0: " + capacity);
 		}
-		int tableSize = 1 << -BitConversion.countLeadingZeros(Math.max(2, (int)Math.ceil(capacity / loadFactor)) - 1);
+		int tableSize = 1 << -BitConversion.countLeadingZeros(Math.max(2, (int)Math.ceil(capacity / (double)loadFactor)) - 1);
 		if (tableSize > 1 << 30 || tableSize < 0) {
 			throw new IllegalArgumentException("The required capacity is too large: " + capacity);
 		}
